refactor: abstract query calls to custom hook
